first draft of geometry library
