+++ src/haddock/modules/analysis/rmsdmatrix/rmsd.py
@@ -5,7 +5,12 @@ from pathlib import Path
 import numpy as np
 
 from haddock import log
-from haddock.libs.libontology import PDBFile
+from haddock.libs.libgeometry import (
+    calc_rmsd,
+    centroid,
+    kabsch,
+    load_coords
+)
 from haddock.modules.analysis.caprieval.capri import get_atoms
 
 
@@ -99,39 +104,9 @@ class RMSD:
         self.output_name = output_name
         self.path = path
         self.atoms = get_atoms(model_list)
-        self.numbering_dic = {}
         # data array
         self.data = np.zeros((self.npairs, 3))
     
-    @staticmethod
-    def calc_rmsd(V, W):
-        """Calculate the RMSD from two vectors."""
-        diff = np.array(V) - np.array(W)
-        N = len(V)
-        return np.sqrt((diff * diff).sum() / N)
-
-    @staticmethod
-    def kabsch(P, Q):
-        """Find the rotation matrix using Kabsch algorithm."""
-        # Covariance matrix
-        P = np.array(P)
-        Q = np.array(Q)
-        C = np.dot(np.transpose(P), Q)
-        # use SVD
-        V, S, W = np.linalg.svd(C)
-        d = (np.linalg.det(V) * np.linalg.det(W)) < 0.0
-        if d:
-            S[-1] = -S[-1]
-            V[:, -1] = -V[:, -1]
-        # Create Rotation matrix U
-        U = np.dot(V, W)
-        return U
-
-    @staticmethod
-    def centroid(X):
-        """Get the centroid."""
-        X = np.array(X)
-        return X.mean(axis=0)
 
     def run(
             self
@@ -143,12 +118,12 @@ class RMSD:
         nmodels = len(self.model_list)
         for n in range(self.npairs):
             
-            ref_coord_dic, _ = self.load_coords(
-                self.model_list[ref], self.filter_resdic, match=False
+            ref_coord_dic, _ = load_coords(
+                self.model_list[ref], self.atoms, self.filter_resdic
                 )
 
-            mod_coord_dic, _ = self.load_coords(
-                self.model_list[mod], self.filter_resdic, match=False
+            mod_coord_dic, _ = load_coords(
+                self.model_list[mod], self.atoms, self.filter_resdic
                 )
             P = []
             Q = []
@@ -159,11 +134,11 @@ class RMSD:
                 P.append(mod_xyz)
             Q = np.asarray(Q)
             P = np.asarray(P)
-            Q = Q - self.centroid(Q)
-            P = P - self.centroid(P)
-            U = self.kabsch(P, Q)
+            Q = Q - centroid(Q)
+            P = P - centroid(P)
+            U = kabsch(P, Q)
             P = np.dot(P, U)
-            rmsd = self.calc_rmsd(P, Q)
+            rmsd = calc_rmsd(P, Q)
             # saving output (adding one for consistency with clusterfcc)
             self.data[n, 0] = ref + 1
             self.data[n, 1] = mod + 1
@@ -175,71 +150,6 @@ class RMSD:
             else:
                 mod += 1
 
-    def load_coords(self, pdb_f, filter_resdic=None, match=False):
-        """Load coordinates from PDB."""
-        coord_dic = {}
-        chain_dic = {}
-        idx = 0
-        if isinstance(pdb_f, PDBFile):
-            pdb_f = pdb_f.rel_path
-        with open(pdb_f, "r") as fh:
-            for line in fh.readlines():
-                if line.startswith("ATOM"):
-
-                    atom_name = line[12:16].strip()
-                    resname = line[17:20].strip()
-                    chain = line[21]
-                    resnum = int(line[22:26])
-
-                    x = float(line[30:38])
-                    y = float(line[38:46])
-                    z = float(line[46:54])
-                    coords = np.asarray([x, y, z])
-
-                    if match:
-                        try:
-                            resnum = self.numbering_dic[chain][resnum]
-                        except KeyError:
-                            # this residue is not matched, and so it should
-                            #  not be considered
-                            # self.log(
-                            #     f"WARNING: {chain}.{resnum}.{atom_name}"
-                            #     " was not matched!"
-                            #     )
-                            continue
-
-                    # identifier = f"{chain}.{resnum}.{atom_name}"
-                    identifier = (chain, resnum, atom_name)
-
-                    if atom_name not in self.atoms[resname]:
-                        continue
-
-                    if chain not in chain_dic:
-                        chain_dic[chain] = []
-
-                    if filter_resdic:
-                        # Only retrieve coordinates from the filter_resdic
-                        if (
-                                chain in filter_resdic
-                                and resnum in filter_resdic[chain]
-                                ):
-                            coord_dic[identifier] = coords
-                            chain_dic[chain].append(idx)
-                            idx += 1
-
-                    else:
-                        # retrieve everything
-                        coord_dic[identifier] = coords
-                        chain_dic[chain].append(idx)
-                        idx += 1
-
-        chain_ranges = {}
-        for chain in chain_dic:
-            min_idx = min(chain_dic[chain])
-            max_idx = max(chain_dic[chain])
-            chain_ranges[chain] = (min_idx, max_idx)
-
-        return coord_dic, chain_ranges
 
     def output(
             self,
