+++ tests/test_module_caprieval.py
@@ -21,11 +21,6 @@ from haddock.modules.analysis.caprieval.capri import (
 from . import golden_data
 
 
-def array_to_list(np_array):
-    """Transform a numpy array in a nested list."""
-    return [list(e) for e in np_array]
-
-
 def round_two_dec(dic):
     """Round a rms dictionary to two digits."""
     return dict((k, round(dic[k], 2)) for k in dic)
@@ -513,77 +508,6 @@ def test_load_contacts(protprot_caprimodule, protprot_input_list):
     assert observed_con_set == expected_con_set
 
 
-def test_calc_rmsd(protprot_caprimodule):
-    """Test the RMSD calculation."""
-    V = [
-        [0.4975639180542826, 2.4634453459517913, 13.502915985158635],
-        [11.227311917422782, 16.269820879072906, -3.962511851240976],
-        [-7.0306634127649135, 4.226046643850031, -8.654935533437529],
-        [-1.592886678968348, -10.874055594404654, -5.820287656251305],
-        [-4.372921399498893, 2.094055858210638, -8.93135311766362],
-        [-1.2876205748591463, -1.4125224802187486, 15.515992917804834],
-        [6.43227627252424, 17.468089367358402, 11.04719510106607],
-        [5.88420286284558, 9.382362089324166, -5.708004059876668],
-        [4.238076687025332, 3.8305358597829944, -5.857279410346993],
-        [5.145346893131023, -2.7233805318567046, 7.910771966100926],
-        ]
-    W = [
-        [-2.2927341772151664, 1.3226734177215107, 12.252610126582304],
-        [8.353265822784834, 10.995673417721513, -7.888389873417699],
-        [-12.224734177215169, -8.316326582278489, 6.479610126582301],
-        [-6.9257341772151655, 6.939673417721508, -1.281389873417698],
-        [-11.438734177215167, -5.325326582278489, 8.2086101265823],
-        [-3.1317341772151686, -1.6973265822784924, 15.408610126582303],
-        [-0.2497341772151671, 15.53367341772151, 4.4986101265822995],
-        [5.965265822784836, 2.219673417721509, -7.501389873417699],
-        [5.648265822784836, -3.0373265822784887, -4.8293898734177],
-        [5.240265822784835, -3.3833265822784924, 9.9936101265823],
-        ]
-
-    rmsd = protprot_caprimodule.calc_rmsd(V, W)
-
-    assert round(rmsd, 2) == 12.02
-
-
-def test_kabsch(protprot_caprimodule):
-    """Test the Kabsch algorithm."""
-    P = [
-        [-3.811237974683542, -0.12069367088607574, 7.200868354430379],
-        [-5.913237974683542, 8.536306329113923, 6.286868354430379],
-        [12.839762025316457, 8.507306329113923, 4.5878683544303795],
-        [4.856762025316458, -3.7666936708860748, -10.57713164556962],
-        [-1.9222379746835419, -4.171693670886076, -6.596131645569622],
-        [-12.186237974683543, -4.807693670886075, -10.77713164556962],
-        [0.3087620253164576, -7.500693670886075, 8.632868354430379],
-        [-0.08023797468354221, -7.218693670886075, 2.9678683544303794],
-        [9.232762025316458, 18.419306329113926, 8.949868354430379],
-        [-1.7482379746835424, -8.395693670886075, 7.604868354430379],
-        ]
-
-    Q = [
-        [6.0952658227848495, -5.630326582278489, 8.033610126582303],
-        [0.5082658227848498, 1.733673417721512, 10.833610126582304],
-        [7.42726582278485, 7.67067341772151, -5.757389873417694],
-        [-12.59373417721515, -6.5263265822784895, 5.217610126582304],
-        [-6.309734177215148, -4.75632658227849, 2.355610126582306],
-        [-3.6647341772151485, -8.990326582278488, -8.014389873417695],
-        [11.789265822784852, -10.62332658227849, 3.966610126582303],
-        [6.895265822784854, -12.33832658227849, 1.2686101265823062],
-        [5.7642658227848536, 16.110673417721507, 1.3446101265823032],
-        [10.306265822784848, -12.226326582278489, 5.028610126582304],
-        ]
-
-    observed_U = protprot_caprimodule.kabsch(P, Q)
-
-    expected_U = [
-        [0.38845040189428726, 0.38160307568742435, -0.8387403518932808],
-        [-0.20700756088693778, 0.9230933741949091, 0.32410876608493205],
-        [0.8979165634978604, 0.047725414019726436, 0.43757071411697324],
-        ]
-
-    np.testing.assert_allclose(np.asarray(expected_U), observed_U)
-
-
 def test_get_atoms():
     """Test the identification of atoms."""
     pdb_list = [
@@ -712,25 +636,6 @@ def test_get_atoms():
     assert observed_atom_dic == expected_atom_dic
 
 
-def test_centroid(protprot_caprimodule):
-    """Test the centroid calculation."""
-    X = [
-        [6.0952658227848495, -5.630326582278489, 8.033610126582303],
-        [0.5082658227848498, 1.733673417721512, 10.833610126582304],
-        [7.42726582278485, 7.67067341772151, -5.757389873417694],
-        ]
-
-    observed_centroid = protprot_caprimodule.centroid(X)
-    observed_centroid = list(observed_centroid)
-    expected_centroid = [
-        4.6769324894515165,
-        1.2580067510548443,
-        4.369943459915638,
-        ]
-
-    assert observed_centroid == expected_centroid
-
-
 def test_add_chain_from_segid(protprot_caprimodule):
     """Test replacing the chainID with segID."""
     tmp = tempfile.NamedTemporaryFile(delete=True)
@@ -743,71 +648,6 @@ def test_add_chain_from_segid(protprot_caprimodule):
         for line in fh:
             if line.startswith("ATOM"):
                 assert line[21] == "A"
-
-
-def test_load_coords(protprot_caprimodule):
-    """Test the loading of coordinates."""
-    # pdb_f = protprot_input_list[0]
-    pdb_f = Path(golden_data, "protein.pdb")
-    (
-        observed_coord_dic,
-        observed_chain_ranges,
-        ) = protprot_caprimodule.load_coords(pdb_f)
-    observed_keys = list(observed_coord_dic.keys())
-    expected_keys = [
-        ("B", 1, "C"),
-        ("B", 1, "O"),
-        ("B", 1, "N"),
-        ("B", 1, "CA"),
-        ("B", 2, "N"),
-        ("B", 2, "CA"),
-        ("B", 2, "C"),
-        ("B", 2, "O"),
-        ("B", 3, "N"),
-        ("B", 3, "CA"),
-        ("B", 3, "C"),
-        ("B", 3, "O"),
-        ("B", 4, "N"),
-        ("B", 4, "CA"),
-        ("B", 4, "C"),
-        ("B", 4, "O"),
-        ("B", 5, "N"),
-        ("B", 5, "CA"),
-        ("B", 5, "C"),
-        ("B", 5, "O"),
-        ]
-
-    assert observed_keys == expected_keys
-
-    observed_coords = array_to_list(observed_coord_dic.values())
-    expected_coords = [
-        [2.76, 8.901, -10.955],
-        [3.081, 10.085, -10.981],
-        [3.315, 8.47, -13.254],
-        [3.439, 7.91, -11.913],
-        [1.846, 8.365, -10.156],
-        [1.091, 9.158, -9.167],
-        [1.355, 8.587, -7.772],
-        [1.393, 7.371, -7.574],
-        [1.625, 9.513, -6.851],
-        [2.045, 9.187, -5.471],
-        [1.203, 9.91, -4.411],
-        [0.519, 10.884, -4.708],
-        [1.149, 9.269, -3.244],
-        [0.62, 9.867, -2.011],
-        [1.447, 9.408, -0.806],
-        [1.613, 8.206, -0.591],
-        [2.009, 10.387, -0.089],
-        [2.761, 10.16, 1.151],
-        [1.832, 10.156, 2.373],
-        [1.352, 11.174, 2.853],
-        ]
-
-    assert observed_coords == expected_coords
-
-    expected_chain_ranges = {"B": (0, 19)}
-
-    assert observed_chain_ranges == expected_chain_ranges
 
 
 def test_pdb2fastadic():
